==> 3_data_from_SMILES_to_text.ipynb <==
Предыдущий файл - 2_dict_SMILES_text.py

#Загрузка данных

import pandas as pd
import json

with open('ended_dict.json', 'r') as file:
    names_dict = json.load(file)

ds = pd.read_csv('prepared_data.csv', sep='\t', index_col = 0)
ds.head()

#Преобразование реакций в текстовый формат

def reaction_to_dict(string):

  """This function converts reaction SMILES into a dictionary of the format {'reactants': [...], 'agents': [...], 'product': [...]}"""

  dict_ = {}
  reactants = string.split('>')[0].split('.')
  dict_['reactants'] = [names_dict[i] if i in names_dict else 'пропуск' for i in reactants]
  products = string.split('>')[-1].split('.')
  dict_['product'] = [names_dict[i] if i in names_dict else 'пропуск'  for i in products]
  if string.split('>')[1] != '':
    agents = string.split('>')[1].split('.')
    dict_['agents'] = [names_dict[i] if i in names_dict else 'пропуск' for i in agents]
  return dict_

def create_sentance(dict_):

  """This function creates a coherent reaction description from the IUPAC names of molecules"""

  reactants = dict_['reactants']
  if len(reactants) > 1:
    reactants = [str(i) for i in reactants]
    reactants_part = ', '.join(reactants)
    part1 = f"{reactants_part} react together"
  else:
    part1 = f'{reactants[0]} reacts'


  if 'agents' in list(dict_.keys()):
    agents = dict_['agents']
    agents = [str(i) for i in agents]
    agents_part = ', '.join(agents)
    part2 = f' in the presence of {agents_part}'
  else:
    part2 = ''
  product_part = dict_['product'][0]
  part3 = f' to produce {product_part}.'
  prompt = part1 + part2 + part3

  return prompt.capitalize()

def create_text_df(df1):

    """This function creates a new dataframe with reaction descriptions instead of SMILES strings"""

    df1['dict'] = df1['ReactionSmiles'].apply(reaction_to_dict)
    df1['text'] = df1['dict'].apply(create_sentance)
    df1 = df1[['text', 'Yield']]
    return df1

df = create_text_df(ds)

counter = 0
for i in ds['text']: #проверка на необработанные молекулы
    if 'пропуск' in i:
        counter += 1
print(f'Количество необработанных молекул: {counter}')

Преобразование таргета

def reaction_class(number):
    return 'high' if number == 1 else 'low'

df['Yield'] = df['Yield'].apply(reaction_class)

Пример готового датасета

pd.set_option('display.max_columns', None)
pd.set_option('display.expand_frame_repr', False)
pd.set_option('display.max_colwidth', None)

df = df.reset_index(drop=True)
df.head()

df.to_csv('text_chem_data.csv', sep = '\t')

Продолжение в 4_train_test_splitting.ipynb
